fase04/Aula 4/Transcricao de Audio/requirements.txt
moviepy
SpeechRecognition
pydub
openai
python-dotenv
yt-dlp
torch
transformers
whisper 
chromadb
tqdm
langchain_community
llama-index-core
llama-index-embeddings-openai
llama-index-vector-stores-chroma
llama-index-llms-langchain